v0.0.7
Finished optical flow motion tracking (but haven't moved the prototype into the main code). It seems to work both quickly and accurately, and it has the added benefit of providing a much more obvious quality tracking signal (i.e. how many of the tracked points remain as the motion develops). The only disadvantage is that it has several new hyperparameters that will need to be tuned. These, however, seem to be far more robust to noise and likely only depend on the scale of the cropped image.